Move out LSPS1-changes
Move out all lsps1-schema's to another file
