+++ libs/gl-client/src/lsps/lsps1/mod.rs
@@ -1,0 +1,1 @@
+pub mod schema;
